Aurum_Data_Fetcher/requirements.txt
requests
python-dotenv
aiohttp
opencc-python-reimplemented
